+++ stream_state_router/__init__.py
@@ -1,3 +1,3 @@
 """Stream State Router."""
 
-__version__ = "2.0.11"
+__version__ = "2.0.12"